Remove match made
