+++ packages/feedback-components/src/feedback/feedback.module.sass
@@ -155,3 +155,11 @@ mark
   border-radius: .2em
   margin-bottom: 4px
   cursor: pointer
+
+.match-container
+  display: flex
+  justify-content: space-between
+  align-items: center
+
+  .close-btn
+    cursor: pointer
